Move TransitMode to modes package
