+++ src/main/java/org/opentripplanner/api/mapping/TraverseModeMapper.java
@@ -1,6 +1,6 @@
 package org.opentripplanner.api.mapping;
 
-import org.opentripplanner.model.TransitMode;
+import org.opentripplanner.model.modes.TransitMode;
 import org.opentripplanner.routing.core.TraverseMode;
 
 import java.util.HashMap;
