+++ src/main/java/org/opentripplanner/graph_builder/module/map/BusRouteStreetMatcher.java
@@ -9,7 +9,7 @@ import org.opentripplanner.model.Route;
 import org.opentripplanner.model.TripPattern;
 import org.opentripplanner.routing.graph.Edge;
 import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.model.TransitMode;
+import org.opentripplanner.model.modes.TransitMode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
