+++ src/main/java/org/opentripplanner/routing/api/request/RoutingRequest.java
@@ -7,7 +7,7 @@ import org.opentripplanner.common.MavenVersion;
 import org.opentripplanner.model.GenericLocation;
 import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.model.Route;
-import org.opentripplanner.model.TransitMode;
+import org.opentripplanner.model.modes.TransitMode;
 import org.opentripplanner.routing.core.IntersectionTraversalCostModel;
 import org.opentripplanner.routing.core.OptimizeType;
 import org.opentripplanner.routing.core.RouteMatcher;
