+++ src/main/java/org/opentripplanner/routing/core/TraverseMode.java
@@ -1,6 +1,6 @@
 package org.opentripplanner.routing.core;
 
-import org.opentripplanner.model.TransitMode;
+import org.opentripplanner.model.modes.TransitMode;
 
 public enum TraverseMode {
     WALK, BICYCLE, CAR,
