+++ src/main/java/org/opentripplanner/util/TravelOptionsMaker.java
@@ -2,7 +2,7 @@ package org.opentripplanner.util;
 
 import org.opentripplanner.routing.core.TraverseMode;
 import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.model.TransitMode;
+import org.opentripplanner.model.modes.TransitMode;
 
 import java.util.ArrayList;
 import java.util.HashSet;
